Add files via upload
This is an example of neural networking (if done with a gpu 50 % faster). 
Note: if using m1, the gpu version will not working since mostly nvidia or amd gpu supported for further integration, but you can use the cpu. 
Tell my ideas on how i can change, later on i will visual the data collected in a graph with the weights and gradients.